requirements.txt
azure-functions
# Additional packages
requests==2.20.1